refactor(flight): split post-switch auto-record predicates into ParsekFlight.PostSwitchPredicates.cs
